Refactor validator accounts exit to remove cli context dependency
