+++ validator/accounts/accounts_exit.go
@@ -9,22 +9,13 @@ import (
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/pkg/errors"
-	grpcutil "github.com/prysmaticlabs/prysm/api/grpc"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
-	"github.com/prysmaticlabs/prysm/cmd"
-	"github.com/prysmaticlabs/prysm/cmd/validator/flags"
 	fieldparams "github.com/prysmaticlabs/prysm/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/config/params"
 	"github.com/prysmaticlabs/prysm/encoding/bytesutil"
-	"github.com/prysmaticlabs/prysm/io/prompt"
 	ethpb "github.com/prysmaticlabs/prysm/proto/prysm/v1alpha1"
-	"github.com/prysmaticlabs/prysm/validator/accounts/iface"
-	"github.com/prysmaticlabs/prysm/validator/accounts/userprompt"
-	"github.com/prysmaticlabs/prysm/validator/accounts/wallet"
 	"github.com/prysmaticlabs/prysm/validator/client"
 	"github.com/prysmaticlabs/prysm/validator/keymanager"
-	"github.com/urfave/cli/v2"
-	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
@@ -37,32 +28,24 @@ type PerformExitCfg struct {
 	FormattedPubKeys []string
 }
 
-const exitPassphrase = "Exit my validator"
+// ExitPassphrase exported for use in test.
+const ExitPassphrase = "Exit my validator"
 
-// ExitAccountsCli performs a voluntary exit on one or more accounts.
-func ExitAccountsCli(cliCtx *cli.Context, r io.Reader) error {
-	validatingPublicKeys, kManager, err := prepareWallet(cliCtx)
-	if err != nil {
-		return err
-	}
-
-	rawPubKeys, trimmedPubKeys, err := interact(cliCtx, r, validatingPublicKeys)
-	if err != nil {
-		return err
-	}
+// Exit performs a voluntary exit on one or more accounts.
+func (acm *AccountsCLIManager) Exit(ctx context.Context, r io.Reader) error {
 	// User decided to cancel the voluntary exit.
-	if rawPubKeys == nil && trimmedPubKeys == nil {
+	if acm.rawPubKeys == nil && acm.formattedPubKeys == nil {
 		return nil
 	}
 
-	validatorClient, nodeClient, err := prepareClients(cliCtx)
+	validatorClient, nodeClient, err := acm.prepareBeaconClients(ctx)
 	if err != nil {
 		return err
 	}
 	if nodeClient == nil {
 		return errors.New("could not prepare beacon node client")
 	}
-	syncStatus, err := (*nodeClient).GetSyncStatus(cliCtx.Context, &emptypb.Empty{})
+	syncStatus, err := (*nodeClient).GetSyncStatus(ctx, &emptypb.Empty{})
 	if err != nil {
 		return err
 	}
@@ -77,11 +60,11 @@ func ExitAccountsCli(cliCtx *cli.Context, r io.Reader) error {
 	cfg := PerformExitCfg{
 		*validatorClient,
 		*nodeClient,
-		kManager,
-		rawPubKeys,
-		trimmedPubKeys,
+		acm.keymanager,
+		acm.rawPubKeys,
+		acm.formattedPubKeys,
 	}
-	rawExitedKeys, trimmedExitedKeys, err := PerformVoluntaryExit(cliCtx.Context, cfg)
+	rawExitedKeys, trimmedExitedKeys, err := PerformVoluntaryExit(ctx, cfg)
 	if err != nil {
 		return err
 	}
@@ -128,115 +111,6 @@ func PerformVoluntaryExit(
 	return rawExitedKeys, formattedExitedKeys, nil
 }
 
-func prepareWallet(cliCtx *cli.Context) (validatingPublicKeys [][fieldparams.BLSPubkeyLength]byte, km keymanager.IKeymanager, err error) {
-	w, err := wallet.OpenWalletOrElseCli(cliCtx, func(cliCtx *cli.Context) (*wallet.Wallet, error) {
-		return nil, wallet.ErrNoWalletFound
-	})
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "could not open wallet")
-	}
-	// TODO(#9883) - Remove this when we have a better way to handle this.
-	if w.KeymanagerKind() == keymanager.Web3Signer {
-		return nil, nil, errors.New(
-			"web3signer wallets cannot exit accounts through cli command yet. please perform this on the remote signer node",
-		)
-	}
-	km, err = w.InitializeKeymanager(cliCtx.Context, iface.InitKeymanagerConfig{ListenForChanges: false})
-	if err != nil {
-		return nil, nil, errors.Wrap(err, ErrCouldNotInitializeKeymanager)
-	}
-	validatingPublicKeys, err = km.FetchValidatingPublicKeys(cliCtx.Context)
-	if err != nil {
-		return nil, nil, err
-	}
-	if len(validatingPublicKeys) == 0 {
-		return nil, nil, errors.New("wallet is empty, no accounts to perform voluntary exit")
-	}
-
-	return validatingPublicKeys, km, nil
-}
-
-func interact(
-	cliCtx *cli.Context,
-	r io.Reader,
-	validatingPublicKeys [][fieldparams.BLSPubkeyLength]byte,
-) (rawPubKeys [][]byte, formattedPubKeys []string, err error) {
-	if !cliCtx.IsSet(flags.ExitAllFlag.Name) {
-		// Allow the user to interactively select the accounts to exit or optionally
-		// provide them via cli flags as a string of comma-separated, hex strings.
-		filteredPubKeys, err := FilterPublicKeysFromUserInput(
-			cliCtx,
-			flags.VoluntaryExitPublicKeysFlag,
-			validatingPublicKeys,
-			userprompt.SelectAccountsVoluntaryExitPromptText,
-		)
-		if err != nil {
-			return nil, nil, errors.Wrap(err, "could not filter public keys for voluntary exit")
-		}
-		rawPubKeys = make([][]byte, len(filteredPubKeys))
-		formattedPubKeys = make([]string, len(filteredPubKeys))
-		for i, pk := range filteredPubKeys {
-			pubKeyBytes := pk.Marshal()
-			rawPubKeys[i] = pubKeyBytes
-			formattedPubKeys[i] = fmt.Sprintf("%#x", bytesutil.Trunc(pubKeyBytes))
-		}
-		allAccountStr := strings.Join(formattedPubKeys, ", ")
-		if !cliCtx.IsSet(flags.VoluntaryExitPublicKeysFlag.Name) {
-			if len(filteredPubKeys) == 1 {
-				promptText := "Are you sure you want to perform a voluntary exit on 1 account? (%s) Y/N"
-				resp, err := prompt.ValidatePrompt(
-					r, fmt.Sprintf(promptText, au.BrightGreen(formattedPubKeys[0])), prompt.ValidateYesOrNo,
-				)
-				if err != nil {
-					return nil, nil, err
-				}
-				if strings.EqualFold(resp, "n") {
-					return nil, nil, nil
-				}
-			} else {
-				promptText := "Are you sure you want to perform a voluntary exit on %d accounts? (%s) Y/N"
-				if len(filteredPubKeys) == len(validatingPublicKeys) {
-					promptText = fmt.Sprintf(
-						"Are you sure you want to perform a voluntary exit on all accounts? Y/N (%s)",
-						au.BrightGreen(allAccountStr))
-				} else {
-					promptText = fmt.Sprintf(promptText, len(filteredPubKeys), au.BrightGreen(allAccountStr))
-				}
-				resp, err := prompt.ValidatePrompt(r, promptText, prompt.ValidateYesOrNo)
-				if err != nil {
-					return nil, nil, err
-				}
-				if strings.EqualFold(resp, "n") {
-					return nil, nil, nil
-				}
-			}
-		}
-	} else {
-		rawPubKeys, formattedPubKeys = prepareAllKeys(validatingPublicKeys)
-		fmt.Printf("About to perform a voluntary exit of %d accounts\n", len(rawPubKeys))
-	}
-
-	promptHeader := au.Red("===============IMPORTANT===============")
-	promptDescription := "Withdrawing funds is not possible in Phase 0 of the system. " +
-		"Please navigate to the following website and make sure you understand the current implications " +
-		"of a voluntary exit before making the final decision:"
-	promptURL := au.Blue("https://docs.prylabs.network/docs/wallet/exiting-a-validator/#withdrawal-delay-warning")
-	promptQuestion := "If you still want to continue with the voluntary exit, please input a phrase found at the end " +
-		"of the page from the above URL"
-	promptText := fmt.Sprintf("%s\n%s\n%s\n%s", promptHeader, promptDescription, promptURL, promptQuestion)
-	resp, err := prompt.ValidatePrompt(r, promptText, func(input string) error {
-		return prompt.ValidatePhrase(input, exitPassphrase)
-	})
-	if err != nil {
-		return nil, nil, err
-	}
-	if strings.EqualFold(resp, "n") {
-		return nil, nil, nil
-	}
-
-	return rawPubKeys, formattedPubKeys, nil
-}
-
 func prepareAllKeys(validatingKeys [][fieldparams.BLSPubkeyLength]byte) (raw [][]byte, formatted []string) {
 	raw = make([][]byte, len(validatingKeys))
 	formatted = make([]string, len(validatingKeys))
@@ -246,29 +120,6 @@ func prepareAllKeys(validatingKeys [][fieldparams.BLSPubkeyLength]byte) (raw [][
 		formatted[i] = fmt.Sprintf("%#x", bytesutil.Trunc(pk[:]))
 	}
 	return
-}
-
-func prepareClients(cliCtx *cli.Context) (*ethpb.BeaconNodeValidatorClient, *ethpb.NodeClient, error) {
-	dialOpts := client.ConstructDialOptions(
-		cliCtx.Int(cmd.GrpcMaxCallRecvMsgSizeFlag.Name),
-		cliCtx.String(flags.CertFlag.Name),
-		cliCtx.Uint(flags.GrpcRetriesFlag.Name),
-		cliCtx.Duration(flags.GrpcRetryDelayFlag.Name),
-	)
-	if dialOpts == nil {
-		return nil, nil, errors.New("failed to construct dial options")
-	}
-
-	grpcHeaders := strings.Split(cliCtx.String(flags.GrpcHeadersFlag.Name), ",")
-	cliCtx.Context = grpcutil.AppendHeaders(cliCtx.Context, grpcHeaders)
-
-	conn, err := grpc.DialContext(cliCtx.Context, cliCtx.String(flags.BeaconRPCProviderFlag.Name), dialOpts...)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "could not dial endpoint %s", flags.BeaconRPCProviderFlag.Name)
-	}
-	validatorClient := ethpb.NewBeaconNodeValidatorClient(conn)
-	nodeClient := ethpb.NewNodeClient(conn)
-	return &validatorClient, &nodeClient, nil
 }
 
 func displayExitInfo(rawExitedKeys [][]byte, trimmedExitedKeys []string) {
